Move validation to its own subpackage
