+++ backend/mora/service/related.py
@@ -17,12 +17,12 @@ This section describes how to interact with related units.
 import flask
 
 from . import handlers
+from .validation import validator
 from .. import common
 from .. import exceptions
 from .. import lora
 from .. import mapping
 from .. import util
-from .. import validator
 
 blueprint = flask.Blueprint('related_unit', __name__, static_url_path='',
                             url_prefix='/service')
