+++ backend/mora/service/role.py
@@ -15,12 +15,12 @@ This section describes how to interact with employee roles.
 '''
 
 from . import handlers
+from .validation import validator
 from .. import common
 from .. import exceptions
 from .. import lora
 from .. import mapping
 from .. import util
-from .. import validator
 
 
 class RoleRequestHandler(handlers.OrgFunkRequestHandler):
