+++ backend/mora/service/validation/validator.py
@@ -7,14 +7,15 @@
 #
 
 import collections
+
 import datetime
 import functools
 import typing
 
-from . import exceptions
-from . import lora
-from . import mapping
-from . import util
+from ... import exceptions
+from ... import lora
+from ... import mapping
+from ... import util
 
 
 def forceable(fn):
